src/third_party/blink/web_tests/http/tests/images/document-policy-unoptimized-lossless-images-strict-expected.txt

CONSOLE ERROR: Document policy violation: unoptimized-lossless-images-strict is not allowed in this document.
CONSOLE ERROR: Document policy violation: unoptimized-lossless-images-strict is not allowed in this document.
